refactor(tui): separate frame localization catalog
