+++ packages/bijou-tui/src/app-frame-i18n.ts
@@ -1,110 +1,15 @@
 import type {
-  I18nCatalog,
   I18nCatalogKey,
   I18nDirection,
   I18nRuntime,
 } from '@flyingrobots/bijou-i18n';
 import type { NotificationHistoryFilter } from './notification.js';
+import { FRAME_I18N_NAMESPACE } from './app-frame-i18n-catalog.js';
 
-const FRAME_I18N_NAMESPACE = 'bijou.shell';
-
-export const FRAME_I18N_CATALOG: I18nCatalog = {
-  namespace: FRAME_I18N_NAMESPACE,
-  entries: [
-    message('key.group.frame', 'Frame'),
-    message('key.group.dock', 'Dock'),
-    message('key.group.scroll', 'Scroll'),
-    message('key.group.shell', 'Shell'),
-    message('key.toggleHelp', 'Toggle help'),
-    message('key.togglePerfHud', 'Toggle perf HUD'),
-    message('key.prevTab', 'Previous tab'),
-    message('key.nextTab', 'Next tab'),
-    message('key.nextPane', 'Next pane'),
-    message('key.prevPane', 'Previous pane'),
-    message('key.search', 'Search'),
-    message('key.openPalette', 'Open command palette'),
-    message('key.toggleThemeMode', 'Toggle theme mode'),
-    message('key.toggleMinimize', 'Fold/unfold pane'),
-    message('key.toggleMaximize', 'Full-screen pane'),
-    message('key.dockUp', 'Move pane up'),
-    message('key.dockDown', 'Move pane down'),
-    message('key.dockLeft', 'Move pane left'),
-    message('key.dockRight', 'Move pane right'),
-    message('key.scrollDown', 'Scroll down'),
-    message('key.scrollUp', 'Scroll up'),
-    message('key.pageDown', 'Page down'),
-    message('key.pageUp', 'Page up'),
-    message('key.top', 'Top'),
-    message('key.bottom', 'Bottom'),
-    message('key.scrollLeft', 'Scroll left'),
-    message('key.scrollRight', 'Scroll right'),
-    message('key.settings', 'Settings'),
-    message('key.notifications', 'Notifications'),
-    message('palette.title', 'Command Palette'),
-    message('palette.hint', 'Enter select • Esc close'),
-    message('search.title', 'Search'),
-    message('perfHud.title', 'Perf HUD'),
-    message('settings.title', 'Settings'),
-    message('settings.footer', 'F2/Esc close • ↑/↓ rows • Enter toggle • / search • q quit'),
-    message('settings.section.shell', 'Shell'),
-    message('settings.shellTheme.label', 'Shell theme'),
-    message('settings.shellTheme.description', 'Current theme: {theme}. Options: {options}.'),
-    message('settings.shellTheme.feedback', 'Shell theme set to {theme}.'),
-    message('settings.shellTheme.modeUnsupported', 'Shell theme {theme} has no alternate mode.'),
-    message('notifications.title', 'Notifications'),
-    message('notifications.footer', 'Shift+N close • f filter • j/k scroll • q quit'),
-    message('notifications.cue.liveArchived', 'notices:{liveCount}+{archivedCount}'),
-    message('notifications.cue.liveOnly', 'notices:{liveCount}'),
-    message('notifications.cue.archivedOnly', 'notices:{archivedCount}'),
-    message('notifications.filter.all', 'All'),
-    message('notifications.filter.actionable', 'Actionable'),
-    message('notifications.summary.liveArchived', 'Live: {liveCount} • Archived: {archivedCount}'),
-    message('notifications.summary.filter', 'Filter: {filter}'),
-    message('notifications.currentStack', 'Current stack'),
-    message('notifications.history.range.empty', '0 items'),
-    message('notifications.history.range.window', '{start}-{end} of {total}'),
-    message('notifications.history.title', 'History • {filter} • {range}'),
-    message('notifications.history.empty', 'No archived notifications for {filter} yet.'),
-    message('notifications.history.action', 'Action: {label}'),
-    message('help.title', 'Keyboard Help'),
-    message('help.hint', 'j/k scroll • d/u page • g/G top/bottom • mouse wheel • ?/Esc close'),
-    message('help.group.general', 'General'),
-    message('help.group.help', 'Help'),
-    message('help.group.settings', 'Settings'),
-    message('help.group.notifications', 'Notifications'),
-    message('help.group.quit', 'Quit'),
-    message('help.key.closeHelp', 'Close help'),
-    message('help.key.closeSettings', 'Close settings'),
-    message('help.key.previousRow', 'Previous row'),
-    message('help.key.nextRow', 'Next row'),
-    message('help.key.activateSetting', 'Activate setting'),
-    message('help.key.closeNotifications', 'Close notification center'),
-    message('help.key.cycleFilter', 'Cycle filter'),
-    message('help.key.quit', 'Quit'),
-    message('help.key.stay', 'Stay'),
-    message('quit.title', 'Quit?'),
-    message('quit.body', 'Quit this app?'),
-    message('quit.footer', 'Y Quit • N Stay'),
-    message('mode.normal', 'NORMAL'),
-    message('mode.palette', 'PALETTE'),
-    message('mode.help', 'HELP'),
-    message('mode.quit', 'QUIT'),
-    message('mode.settings', 'SETTINGS'),
-    message('mode.notices', 'NOTICES'),
-    message('mode.modal', 'MODAL'),
-  ],
-};
-
-function message(id: string, value: string) {
-  return {
-    key: { namespace: FRAME_I18N_NAMESPACE, id },
-    kind: 'message' as const,
-    sourceLocale: 'en',
-    values: { en: value },
-  };
-}
-
-function interpolate(template: string, values: Readonly<Record<string, unknown>>): string {
+function interpolate(
+  template: string,
+  values: Readonly<Record<string, unknown>>,
+): string {
   return template.replace(/\{([^}]+)\}/g, (_match, rawKey: string) => {
     const value = values[rawKey];
     switch (typeof value) {
@@ -151,17 +56,22 @@ export function frameDirection(i18n: I18nRuntime | undefined): I18nDirection {
   return i18n?.direction ?? 'ltr';
 }
 
-export function frameStartAnchor(i18n: I18nRuntime | undefined): 'left' | 'right' {
+export function frameStartAnchor(
+  i18n: I18nRuntime | undefined,
+): 'left' | 'right' {
   return frameDirection(i18n) === 'rtl' ? 'right' : 'left';
 }
 
-export function frameEndAnchor(i18n: I18nRuntime | undefined): 'left' | 'right' {
+export function frameEndAnchor(
+  i18n: I18nRuntime | undefined,
+): 'left' | 'right' {
   return frameDirection(i18n) === 'rtl' ? 'left' : 'right';
 }
 
 export function frameModeLabel(
   i18n: I18nRuntime | undefined,
-  mode: 'NORMAL' | 'PALETTE' | 'HELP' | 'QUIT' | 'SETTINGS' | 'NOTICES' | 'MODAL',
+  mode:
+    'NORMAL' | 'PALETTE' | 'HELP' | 'QUIT' | 'SETTINGS' | 'NOTICES' | 'MODAL',
 ): string {
   return frameMessage(i18n, `mode.${mode.toLowerCase()}`, mode);
 }
@@ -193,10 +103,22 @@ export function frameNotificationCue(
     );
   }
   if (liveCount > 0) {
-    return frameMessage(i18n, 'notifications.cue.liveOnly', 'notices:{liveCount}', { liveCount });
+    return frameMessage(
+      i18n,
+      'notifications.cue.liveOnly',
+      'notices:{liveCount}',
+      { liveCount },
+    );
   }
   if (archivedCount > 0) {
-    return frameMessage(i18n, 'notifications.cue.archivedOnly', 'notices:{archivedCount}', { archivedCount });
+    return frameMessage(
+      i18n,
+      'notifications.cue.archivedOnly',
+      'notices:{archivedCount}',
+      { archivedCount },
+    );
   }
   return undefined;
 }
+
+export { FRAME_I18N_CATALOG } from './app-frame-i18n-catalog.js';
